move duplicated functions to utils
